+++ src/workflows/rna/log_normalize/integration_test.sh
@@ -8,17 +8,8 @@ cd "$REPO_ROOT"
 
 nextflow \
   run . \
-  -main-script src/workflows/integration/scvi_leiden/test.nf \
+  -main-script src/workflows/rna/log_normalize/test.nf \
   -entry test_wf \
-  -profile docker,no_publish \
-  -c src/workflows/utils/labels_ci.config \
-  -c src/workflows/utils/integration_tests.config
-
-
-nextflow \
-  run . \
-  -main-script src/workflows/integration/scvi_leiden/test.nf \
-  -entry test_tiledb_wf \
   -profile docker,no_publish \
   -c src/workflows/utils/labels_ci.config \
   -c src/workflows/utils/integration_tests.config